Refactor: Merge worker package into concurrent
This commit merges the `pkg/common/worker` package into `pkg/common/concurrent` to consolidate related concurrency utilities.

The `Pool` struct has been renamed to `WorkerPool` for better clarity and to avoid ambiguity with other potential pool types. All import paths and references have been updated accordingly throughout the project.
